v0.2.0 — add face detection, text recognition, pose detection
New modules: mlkit-face-detection, mlkit-text-recognition,
mlkit-pose-detection, and their Play Services variants.

Wiki updated with developer guides for all new features,
updated Setup page with all v0.2.0 artifacts, updated
Roadmap and Sidebar, added .gitignore and CLAUDE.md.

+++ barcode-play-services/src/iosMain/kotlin/io/github/rufenkhokhar/barcode/IOSBarcodeProcessor.kt
@@ -116,8 +116,8 @@ internal class IOSBarcodeProcessor(
         val cornerPoints = barcode.cornerPoints?.mapNotNull { value ->
             (value as? NSValue)?.CGPointValue()?.useContents {
                     io.github.rufenkhokhar.corevision.Point(
-                        x = x.toInt(),
-                        y = y.toInt()
+                        x = x.toFloat(),
+                        y = y.toFloat()
                     )
                 }
         }
